+++ R/get_fitvals.R
@@ -1,9 +1,9 @@
-#' @title get_fitvals: extract fit values from glm fit object
+#' @title glm_fitvals: extract fit values from glm fit object
 #' 
 #' @description Helper function which extracts needed parameters from glm 
 #' fit object  
 #'
-#' @usage get_fitvals(glm.fit, data, alpha=0.05)
+#' @usage glm_fitvals(glm.fit, data, alpha=0.05)
 #' 
 #' @param alpha The threshold for CI interval, default is 0.05 (one-sided).  
 #' 
@@ -11,7 +11,7 @@
 #' 
 #' @export
 #' 
-get_fitvals <- function(glm.fit, data, alpha=0.05) {
+glm_fitvals <- function(glm.fit, data, alpha=0.05) {
   
   pred <- predict(glm.fit, 
                   newdata=data, 
@@ -23,6 +23,36 @@ get_fitvals <- function(glm.fit, data, alpha=0.05) {
   lower <- pred$fit - (qnorm(1-alpha) * pred$se.fit)
   upper.fit <- glm.fit$family$linkinv(upper)
   lower.fit <- glm.fit$family$linkinv(lower)
+  
+  res <- list(fitted = fitted, 
+              upper.fit = upper.fit,
+              lower.fit = lower.fit)
+  return(res)
+}
+
+#' @title gam_fitvals: extract fit values from glm fit object
+#' 
+#' @description Helper function which extracts needed parameters from gam 
+#' fit object  
+#'
+#' @usage gam_fitvals(gam.fit, data, alpha=0.05)
+#' 
+#' @param alpha The threshold for CI interval, default is 0.05 (one-sided).  
+#' 
+#' @export
+#' 
+gam_fitvals <- function(gam.fit, data, alpha=0.05) {
+  
+  pred <- predict(gam.fit, 
+                  newdata=data, 
+                  se.fit=TRUE, 
+                  type="link")
+  
+  fitted <- gam.fit$family$linkinv(pred$fit)
+  upper <- pred$fit + (qnorm(1-alpha) * pred$se.fit)
+  lower <- pred$fit - (qnorm(1-alpha) * pred$se.fit)
+  upper.fit <- gam.fit$family$linkinv(upper)
+  lower.fit <- gam.fit$family$linkinv(lower)
   
   res <- list(fitted = fitted, 
               upper.fit = upper.fit,