+++ data-raw/mk_simdata.R
@@ -1,0 +1,5 @@
+##
+##  Produce a simulated dataset for testing assumptions
+##  
+##
+##