+++ tests/testthat/test_03_fluserf.R
@@ -11,7 +11,8 @@ test_that("fluserf computes correct estimates", {
 
   #load data
   dnames <- c("year", "week", "fludeaths", "alldeaths", 
-              "perc_fludeaths", "yrweek_dt", "prop_flupos", "week_in_order", 
+              "perc_fludeaths", "yrweek_dt", "fluyr",
+              "prop_flupos", "week_in_order", 
               "epi", "y0", "y0_ul")
   
   #Tests