+++ tests/testthat/test_05_fludiff.R
@@ -9,7 +9,8 @@ test_that("fludiff output is canon", {
   
   #load data
   dnames <- c("year", "week", "fludeaths", "alldeaths", 
-              "perc_fludeaths", "yrweek_dt", "prop_flupos", "week_in_order",
+              "perc_fludeaths", "yrweek_dt", "fluyr", 
+              "prop_flupos", "week_in_order",
               "epi", "y0", "y0_ul", "y_diff")
   
   #Tests